requirements.txt
cirq
cirq_google
qsimcirq
qiskit
qiskit_aer
qiskit-ibm-runtime
ply
pyyaml
numpy
window-curses ; platform_system == 'Windows'
pydantic>=2.5.0

## Sort out dependency problem with pydantic 2.5.0 for qiskit and smaller than 2.0.0 for cirq.